pyproject.toml
[project]
name = "project-recipe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.9.2",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.127.0",
    "motor>=3.7.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-jose>=3.5.0",
    "redis>=7.1.0",
    "tortoise-orm>=0.25.2",
]

[tool.aerich]
tortoise_orm = "app.db.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
